uint8

+++ midi_controller/leds.py
@@ -39,7 +39,7 @@ class LedInterface(object):
         self.buffer = np.zeros((length, 3))
 
         # allocate new array
-        self.raw_buffer = (ctypes.c_uint32 * self.length)()
+        self.raw_buffer = (ctypes.c_uint8 * self.length)()
         # ws.ws2811_channel_t_leds_set(self._hw._channel, ctypes.byref(self.raw_buffer))
         ws.set_buffer(self._hw._channel, ctypes.byref(self.raw_buffer))
